+++ public/js/editor-common.js
@@ -63,47 +63,47 @@ function thumbnailManager(initial = {}) {
     }
 }
 
-window.gridIsMultiple = false;
-window.openGridModal = function(
-    context,
-    multiple = window.gridIsMultiple,
-    selected = [],
-    labels = {},
-    details = {}
-) {
+// window.gridIsMultiple = false;
+// window.openGridModal = function(
+//     context,
+//     multiple = window.gridIsMultiple,
+//     selected = [],
+//     labels = {},
+//     details = {}
+// ) {
 
-    labels = {
-        ...{
-            title: 'Selecionar Mídia',
-            insert: 'Inserir',
-            bulkInsert: 'Inserir selecionados',
-            clear: 'Limpar',
-            close: 'Fechar'
-        },
-        ...labels
-    };
-    window.dispatchEvent(new CustomEvent('media:set-labels', { detail: { labels } }));
-    window.dispatchEvent(new CustomEvent('modal-set-title', {
-        detail: { id: 'selectorModal', title: labels.title }
-    }));
-    window.dispatchEvent(new CustomEvent('modal-set-close-label', {
-        detail: { id: 'selectorModal', closeLabel: labels.close }
-    }));
+//     labels = {
+//         ...{
+//             title: 'Selecionar Mídia',
+//             insert: 'Inserir',
+//             bulkInsert: 'Inserir selecionados',
+//             clear: 'Limpar',
+//             close: 'Fechar'
+//         },
+//         ...labels
+//     };
+//     window.dispatchEvent(new CustomEvent('media:set-labels', { detail: { labels } }));
+//     window.dispatchEvent(new CustomEvent('modal-set-title', {
+//         detail: { id: 'selectorModal', title: labels.title }
+//     }));
+//     window.dispatchEvent(new CustomEvent('modal-set-close-label', {
+//         detail: { id: 'selectorModal', closeLabel: labels.close }
+//     }));
 
-    // Guarda a última opção ao abrir o modal
-    window.gridIsMultiple = multiple;
+//     // Guarda a última opção ao abrir o modal
+//     window.gridIsMultiple = multiple;
 
-    // Configura o grid para aceitar ou não seleção múltipla
-    window.dispatchEvent(new CustomEvent('media:set-multiple', { detail: { multiple } }));
+//     // Configura o grid para aceitar ou não seleção múltipla
+//     window.dispatchEvent(new CustomEvent('media:set-multiple', { detail: { multiple } }));
 
-    // Configura o grid para mostrar imagens já seleionadas
-    window.dispatchEvent(new CustomEvent('media:set-selected', { detail: { selected } }));
+//     // Configura o grid para mostrar imagens já seleionadas
+//     window.dispatchEvent(new CustomEvent('media:set-selected', { detail: { selected } }));
 
-    // Abre o modal padrão do seletor
-    window.dispatchEvent(new CustomEvent('modal-open', {
-        detail: { id: 'selectorModal', context, ...details }
-    }));
-};
+//     // Abre o modal padrão do seletor
+//     window.dispatchEvent(new CustomEvent('modal-open', {
+//         detail: { id: 'selectorModal', context, ...details }
+//     }));
+// };
 
 // Eventos globais de Thumbnail e Upload
 window.addEventListener('media:uploaded', (e) => {